template_http/requirements.txt
google-cloud-storage==1.33.0
flask